requirements.txt
crawl4ai
PyYAML>=6.0.0
zhipuai
beautifulsoup4
